+++ src/index.js
@@ -1,4 +1,8 @@
 export default function setIn(obj, path, updater) {
+  if (!Array.isArray(path)) {
+    throw new Error('path must be an array');
+  }
+
   if (!path.length) {
     return updater(obj);
   }
@@ -6,38 +10,64 @@ export default function setIn(obj, path, updater) {
   path = path.slice();
 
   const accessor = path.shift();
-  let value = typeof obj !== 'undefined' && obj[accessor];
+  const value = typeof obj !== 'undefined' && obj[accessor];
+  let nextObj = obj;
 
-  if (typeof accessor === 'string' && (typeof obj !== 'object' || Array.isArray(obj))) {
-    obj = {};
-  } else if (typeof accessor === 'number' && !Array.isArray(obj)) {
-    obj = [];
+  if (typeof accessor === 'string' && (typeof nextObj !== 'object' || Array.isArray(nextObj))) {
+    nextObj = {};
+  } else if (typeof accessor === 'number' && !Array.isArray(nextObj)) {
+    nextObj = [];
   }
 
   if (typeof accessor === 'number') {
-    obj = obj && obj.slice();
-
     if (updater || path.length) {
       if (accessor === -1) {
-        obj.push(updater());
-      } else {
-        obj[accessor] = setIn(value, path, updater);
+        return [...nextObj, setIn([], path, updater)];
       }
-    } else {
-      obj.splice(accessor, 1);
+
+      const nextValue = setIn(value, path, updater);
+
+      if (typeof nextValue !== 'undefined') {
+        if (nextValue === value) {
+          return obj;
+        } else {
+          nextObj = [...nextObj];
+          nextObj[accessor] = nextValue;
+
+          return nextObj;
+        }
+      }
     }
 
-    return obj;
+    // If updater returned undefined or no updater at all, delete the item
+    if (accessor in nextObj) {
+      nextObj = [...nextObj];
+      nextObj.splice(accessor, 1);
+    }
+
+    return nextObj;
   } else {
     if (updater || path.length) {
-      return {
-        ...obj,
-        [accessor]: setIn(value, path, updater)
-      };
-    } else {
-      const { [accessor]: deleted, ...nextObj } = obj;
+      const nextValue = setIn(value, path, updater);
 
-      return nextObj;
+      if (typeof nextValue !== 'undefined') {
+        if (nextValue === value) {
+          return obj;
+        } else {
+          return {
+            ...nextObj,
+            [accessor]: nextValue
+          };
+        }
+      }
     }
+
+    // If updater returned undefined or no updater at all, delete the key
+    if (accessor in nextObj) {
+      nextObj = { ...nextObj };
+      delete nextObj[accessor];
+    }
+
+    return nextObj;
   }
 }